feat: load project specific clang-format versions.
This makes it easier for projects to enforce formatting with a specific
clang-format version.

Uses node-resolve to find a clang-format relative to (in order) the last
given file to format, then the current working directory. After that,
falls back to the clang-format that came with this node module.

This is not using Liftoff; it makes too many assumptions on the project
type, e.g. looking for a configuration file a la `gulpfile.js`.