fix: check-clang-format not working on windows
Currently when someone runs the "check-clang-format"
binary from the "clang-format" package on Windows, the
script will always fail. This is because on Windows, the
shebang cannot be read, and only the "PATHEXT" environment
variable is respected. Therefore we can just explicitly
declare that we want to run the "python" executable in the
"PATH" environment variable. This ensures that the Python
script can be launched properly on windows.